+++ python_client/chip/clusters/cluster_defs/Thermostat.py
@@ -80,6 +80,10 @@ class Thermostat(Cluster):
                 ClusterObjectFieldDescriptor(Label="presets", Tag=0x00000050, Type=typing.Optional[typing.List[Thermostat.Structs.PresetStruct]]),
                 ClusterObjectFieldDescriptor(Label="schedules", Tag=0x00000051, Type=typing.Optional[typing.List[Thermostat.Structs.ScheduleStruct]]),
                 ClusterObjectFieldDescriptor(Label="setpointHoldExpiryTimestamp", Tag=0x00000052, Type=typing.Union[None, Nullable, uint]),
+                ClusterObjectFieldDescriptor(Label="maxThermostatSuggestions", Tag=0x00000053, Type=typing.Optional[uint]),
+                ClusterObjectFieldDescriptor(Label="thermostatSuggestions", Tag=0x00000054, Type=typing.Optional[typing.List[Thermostat.Structs.ThermostatSuggestionStruct]]),
+                ClusterObjectFieldDescriptor(Label="currentThermostatSuggestion", Tag=0x00000055, Type=typing.Union[None, Nullable, Thermostat.Structs.ThermostatSuggestionStruct]),
+                ClusterObjectFieldDescriptor(Label="thermostatSuggestionNotFollowingReason", Tag=0x00000056, Type=typing.Union[None, Nullable, Thermostat.Bitmaps.ThermostatSuggestionNotFollowingReasonBitmap]),
                 ClusterObjectFieldDescriptor(Label="generatedCommandList", Tag=0x0000FFF8, Type=typing.List[uint]),
                 ClusterObjectFieldDescriptor(Label="acceptedCommandList", Tag=0x0000FFF9, Type=typing.List[uint]),
                 ClusterObjectFieldDescriptor(Label="eventList", Tag=0x0000FFFA, Type=typing.List[uint]),
@@ -145,6 +149,10 @@ class Thermostat(Cluster):
     presets: typing.Optional[typing.List[Thermostat.Structs.PresetStruct]] = None
     schedules: typing.Optional[typing.List[Thermostat.Structs.ScheduleStruct]] = None
     setpointHoldExpiryTimestamp: typing.Union[None, Nullable, uint] = None
+    maxThermostatSuggestions: typing.Optional[uint] = None
+    thermostatSuggestions: typing.Optional[typing.List[Thermostat.Structs.ThermostatSuggestionStruct]] = None
+    currentThermostatSuggestion: typing.Union[None, Nullable, Thermostat.Structs.ThermostatSuggestionStruct] = None
+    thermostatSuggestionNotFollowingReason: typing.Union[None, Nullable, Thermostat.Bitmaps.ThermostatSuggestionNotFollowingReasonBitmap] = None
     generatedCommandList: typing.List[uint] = field(default_factory=lambda: [])
     acceptedCommandList: typing.List[uint] = field(default_factory=lambda: [])
     eventList: typing.List[uint] = field(default_factory=lambda: [])
@@ -313,6 +321,8 @@ class Thermostat(Cluster):
             kLocalTemperatureNotExposed = 0x40
             kMatterScheduleConfiguration = 0x80
             kPresets = 0x100
+            kEvents = 0x200
+            kThermostatSuggestions = 0x400
 
         class ACErrorCodeBitmap(IntFlag):
             kCompressorFail = 0x1
@@ -361,6 +371,16 @@ class Thermostat(Cluster):
         class ScheduleModeBitmap(IntFlag):
             kHeatSetpointPresent = 0x1
             kCoolSetpointPresent = 0x2
+
+        class ThermostatSuggestionNotFollowingReasonBitmap(IntFlag):
+            kDemandResponseEvent = 0x1
+            kOngoingHold = 0x2
+            kSchedule = 0x4
+            kOccupancy = 0x8
+            kVacationMode = 0x10
+            kTimeOfUseCostSavings = 0x20
+            kPreCoolingOrPreHeating = 0x40
+            kConflictingSuggestions = 0x80
 
     class Structs:
         @dataclass
@@ -471,6 +491,23 @@ class Thermostat(Cluster):
             numberOfSchedules: uint = 0
             scheduleTypeFeatures: Thermostat.Bitmaps.ScheduleTypeFeaturesBitmap = 0
 
+        @dataclass
+        class ThermostatSuggestionStruct(ClusterObject):
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="uniqueID", Tag=0, Type=uint),
+                        ClusterObjectFieldDescriptor(Label="presetHandle", Tag=1, Type=bytes),
+                        ClusterObjectFieldDescriptor(Label="effectiveTime", Tag=2, Type=uint),
+                        ClusterObjectFieldDescriptor(Label="expirationTime", Tag=3, Type=uint),
+                    ])
+
+            uniqueID: uint = 0
+            presetHandle: bytes = b""
+            effectiveTime: uint = 0
+            expirationTime: uint = 0
+
     class Commands:
         @dataclass
         class SetpointRaiseLower(ClusterCommand):
@@ -523,6 +560,42 @@ class Thermostat(Cluster):
             presetHandle: typing.Union[Nullable, bytes] = NullValue
 
         @dataclass
+        class AddThermostatSuggestion(ClusterCommand):
+            cluster_id: typing.ClassVar[int] = 0x00000201
+            command_id: typing.ClassVar[int] = 0x00000007
+            is_client: typing.ClassVar[bool] = True
+            response_type: typing.ClassVar[typing.Optional[str]] = 'AddThermostatSuggestionResponse'
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="presetHandle", Tag=0, Type=bytes),
+                        ClusterObjectFieldDescriptor(Label="effectiveTime", Tag=1, Type=typing.Union[Nullable, uint]),
+                        ClusterObjectFieldDescriptor(Label="expirationInMinutes", Tag=2, Type=uint),
+                    ])
+
+            presetHandle: bytes = b""
+            effectiveTime: typing.Union[Nullable, uint] = NullValue
+            expirationInMinutes: uint = 0
+
+        @dataclass
+        class RemoveThermostatSuggestion(ClusterCommand):
+            cluster_id: typing.ClassVar[int] = 0x00000201
+            command_id: typing.ClassVar[int] = 0x00000008
+            is_client: typing.ClassVar[bool] = True
+            response_type: typing.ClassVar[typing.Optional[str]] = None
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="uniqueID", Tag=0, Type=uint),
+                    ])
+
+            uniqueID: uint = 0
+
+        @dataclass
         class AtomicRequest(ClusterCommand):
             cluster_id: typing.ClassVar[int] = 0x00000201
             command_id: typing.ClassVar[int] = 0x000000FE
@@ -541,6 +614,22 @@ class Thermostat(Cluster):
             requestType: Globals.Enums.enum8 = 0
             attributeRequests: typing.List[uint] = field(default_factory=lambda: [])
             timeout: typing.Optional[uint] = None
+
+        @dataclass
+        class AddThermostatSuggestionResponse(ClusterCommand):
+            cluster_id: typing.ClassVar[int] = 0x00000201
+            command_id: typing.ClassVar[int] = 0x00000002
+            is_client: typing.ClassVar[bool] = False
+            response_type: typing.ClassVar[typing.Optional[str]] = None
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="uniqueID", Tag=0, Type=uint),
+                    ])
+
+            uniqueID: uint = 0
 
         @dataclass
         class AtomicResponse(ClusterCommand):
@@ -1476,6 +1565,70 @@ class Thermostat(Cluster):
             value: typing.Union[None, Nullable, uint] = None
 
         @dataclass
+        class MaxThermostatSuggestions(ClusterAttributeDescriptor):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def attribute_id(cls) -> int:
+                return 0x00000053
+
+            @ChipUtility.classproperty
+            def attribute_type(cls) -> ClusterObjectFieldDescriptor:
+                return ClusterObjectFieldDescriptor(Type=typing.Optional[uint])
+
+            value: typing.Optional[uint] = None
+
+        @dataclass
+        class ThermostatSuggestions(ClusterAttributeDescriptor):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def attribute_id(cls) -> int:
+                return 0x00000054
+
+            @ChipUtility.classproperty
+            def attribute_type(cls) -> ClusterObjectFieldDescriptor:
+                return ClusterObjectFieldDescriptor(Type=typing.Optional[typing.List[Thermostat.Structs.ThermostatSuggestionStruct]])
+
+            value: typing.Optional[typing.List[Thermostat.Structs.ThermostatSuggestionStruct]] = None
+
+        @dataclass
+        class CurrentThermostatSuggestion(ClusterAttributeDescriptor):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def attribute_id(cls) -> int:
+                return 0x00000055
+
+            @ChipUtility.classproperty
+            def attribute_type(cls) -> ClusterObjectFieldDescriptor:
+                return ClusterObjectFieldDescriptor(Type=typing.Union[None, Nullable, Thermostat.Structs.ThermostatSuggestionStruct])
+
+            value: typing.Union[None, Nullable, Thermostat.Structs.ThermostatSuggestionStruct] = None
+
+        @dataclass
+        class ThermostatSuggestionNotFollowingReason(ClusterAttributeDescriptor):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def attribute_id(cls) -> int:
+                return 0x00000056
+
+            @ChipUtility.classproperty
+            def attribute_type(cls) -> ClusterObjectFieldDescriptor:
+                return ClusterObjectFieldDescriptor(Type=typing.Union[None, Nullable, Thermostat.Bitmaps.ThermostatSuggestionNotFollowingReasonBitmap])
+
+            value: typing.Union[None, Nullable, Thermostat.Bitmaps.ThermostatSuggestionNotFollowingReasonBitmap] = None
+
+        @dataclass
         class GeneratedCommandList(ClusterAttributeDescriptor):
             @ChipUtility.classproperty
             def cluster_id(cls) -> int:
@@ -1570,3 +1723,174 @@ class Thermostat(Cluster):
                 return ClusterObjectFieldDescriptor(Type=uint)
 
             value: uint = 0
+
+    class Events:
+        @dataclass
+        class SystemModeChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000000
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="previousSystemMode", Tag=0, Type=typing.Optional[Thermostat.Enums.SystemModeEnum]),
+                        ClusterObjectFieldDescriptor(Label="currentSystemMode", Tag=1, Type=Thermostat.Enums.SystemModeEnum),
+                    ])
+
+            previousSystemMode: typing.Optional[Thermostat.Enums.SystemModeEnum] = None
+            currentSystemMode: Thermostat.Enums.SystemModeEnum = 0
+
+        @dataclass
+        class LocalTemperatureChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000001
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="currentLocalTemperature", Tag=0, Type=typing.Union[Nullable, int]),
+                    ])
+
+            currentLocalTemperature: typing.Union[Nullable, int] = NullValue
+
+        @dataclass
+        class OccupancyChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000002
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="previousOccupancy", Tag=0, Type=typing.Optional[Thermostat.Bitmaps.OccupancyBitmap]),
+                        ClusterObjectFieldDescriptor(Label="currentOccupancy", Tag=1, Type=Thermostat.Bitmaps.OccupancyBitmap),
+                    ])
+
+            previousOccupancy: typing.Optional[Thermostat.Bitmaps.OccupancyBitmap] = None
+            currentOccupancy: Thermostat.Bitmaps.OccupancyBitmap = 0
+
+        @dataclass
+        class SetpointChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000003
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="systemMode", Tag=0, Type=Thermostat.Enums.SystemModeEnum),
+                        ClusterObjectFieldDescriptor(Label="occupancy", Tag=1, Type=typing.Optional[Thermostat.Bitmaps.OccupancyBitmap]),
+                        ClusterObjectFieldDescriptor(Label="previousSetpoint", Tag=2, Type=typing.Optional[int]),
+                        ClusterObjectFieldDescriptor(Label="currentSetpoint", Tag=3, Type=int),
+                    ])
+
+            systemMode: Thermostat.Enums.SystemModeEnum = 0
+            occupancy: typing.Optional[Thermostat.Bitmaps.OccupancyBitmap] = None
+            previousSetpoint: typing.Optional[int] = None
+            currentSetpoint: int = 0
+
+        @dataclass
+        class RunningStateChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000004
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="previousRunningState", Tag=0, Type=typing.Optional[Thermostat.Bitmaps.RelayStateBitmap]),
+                        ClusterObjectFieldDescriptor(Label="currentRunningState", Tag=1, Type=Thermostat.Bitmaps.RelayStateBitmap),
+                    ])
+
+            previousRunningState: typing.Optional[Thermostat.Bitmaps.RelayStateBitmap] = None
+            currentRunningState: Thermostat.Bitmaps.RelayStateBitmap = 0
+
+        @dataclass
+        class RunningModeChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000005
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="previousRunningMode", Tag=0, Type=typing.Optional[Thermostat.Enums.ThermostatRunningModeEnum]),
+                        ClusterObjectFieldDescriptor(Label="currentRunningMode", Tag=1, Type=Thermostat.Enums.ThermostatRunningModeEnum),
+                    ])
+
+            previousRunningMode: typing.Optional[Thermostat.Enums.ThermostatRunningModeEnum] = None
+            currentRunningMode: Thermostat.Enums.ThermostatRunningModeEnum = 0
+
+        @dataclass
+        class ActiveScheduleChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000006
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="previousScheduleHandle", Tag=0, Type=typing.Union[None, Nullable, bytes]),
+                        ClusterObjectFieldDescriptor(Label="currentScheduleHandle", Tag=1, Type=typing.Union[Nullable, bytes]),
+                    ])
+
+            previousScheduleHandle: typing.Union[None, Nullable, bytes] = None
+            currentScheduleHandle: typing.Union[Nullable, bytes] = NullValue
+
+        @dataclass
+        class ActivePresetChange(ClusterEvent):
+            @ChipUtility.classproperty
+            def cluster_id(cls) -> int:
+                return 0x00000201
+
+            @ChipUtility.classproperty
+            def event_id(cls) -> int:
+                return 0x00000007
+
+            @ChipUtility.classproperty
+            def descriptor(cls) -> ClusterObjectDescriptor:
+                return ClusterObjectDescriptor(
+                    Fields=[
+                        ClusterObjectFieldDescriptor(Label="previousPresetHandle", Tag=0, Type=typing.Union[None, Nullable, bytes]),
+                        ClusterObjectFieldDescriptor(Label="currentPresetHandle", Tag=1, Type=typing.Union[Nullable, bytes]),
+                    ])
+
+            previousPresetHandle: typing.Union[None, Nullable, bytes] = None
+            currentPresetHandle: typing.Union[Nullable, bytes] = NullValue